use the LU decomposition if Cholesky fails
this actually works in the situations I was encountering for rust-semp, so I'll
leave it in. I think the Cholesky decomposition is preferable for efficiency and
accuracy if it works, but in lieu of that, the LU version should suffice. If the
LU decomposition fails, it still panics as well.